Add trailing period to domain.hostname_reversed.
This will allow us to more easily query by domains down the line by preventing name collisions.

For example, if we want the top-level domain and all the sub domains of fcc.gov, we could search for gov.fcc.

However, if there also was fcclabs.gov, a search for gov.fcc% would return both, and a search for gov.fcc. would not return the primary domain.

To avoid this, add a trailing period to all domain names, such that we can query for gov.fcc.% and get both the top-level domain, and subdomains more easily.

This commit both updates the model and includes a bi-directional migration.